[REF] client: consolidate protocol contracts
