+++ crates/client/test/support/real_protocol_core.mjs
@@ -4,7 +4,7 @@ import init, { ProtocolCoreWasm } from "../../generated/o_sfu_protocol.js";
 import {
     configureDefaultWasmProtocolCoreProvider,
     createProtocolCore
-} from "../../dist/runtime_contract.js";
+} from "../../dist/protocol_contract.js";
 
 await init({
     module_or_path: await readFile(
